requirements.txt
Django==1.7.5
Fabric==1.10.1
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
django-ipware==0.1.0
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==3.0.5
drf-nested-routers==0.9.0
ecdsa==0.13
gunicorn==19.3.0
paramiko==1.15.2
psycopg2==2.6
pycrypto==2.6.1
pytz==2015.2
static3==0.5.1
wsgiref==0.1.2
